rpatchur: Bump web-view version to 0.7.3
Update CHANGELOG
Fix several clippy warnings

+++ gruf/src/thor/builder.rs
@@ -272,7 +272,7 @@ where
     W: Write,
 {
     // Use an 8KiB buffer
-    let mut buf = [0 as u8; 8 * 1024];
+    let mut buf = [0_u8; 8 * 1024];
     let mut digest = crc32::Digest::new(crc::crc32::IEEE);
     let mut written = 0;
     loop {
